requirements.txt
fastapi
uvicorn
dnspython
requests
httpx